+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.3.0
 
-- Only `Bundler.setup(:default)` when using as an API to avoid loading :development/:test group gems into a production app 
+- Setup Bundler default group only when using as an API to avoid loading :development/:test group gems into a production app 
+- Announce that download files are missing for `bundler download show` when a Downloadfile exists but downloads have not been made
 
 ## 1.2.0